#3 New Implementation of Borda
A new approach for borda has been implemented, it seems to work, as it gives the same results as the example on the board (which is a good sign given it is the same exact scenario). the documentation might need to be updated in order to give a better explanation of the borda voting rule.