Most recent updates
The code has been refactored and updated so it can handle candidates that are stadiums (via csv input). Additionally, these candidates are not given x and y coordinates based on the stadium zipcodes, and the stadiums are all plotted on a graph. The voting algorithms have all been commented out mainly because there is currently no data for voters, so it would cause a crash to run the code with the voting strategies. Once the voter data is in, then there will be a function that needs to be made that can rank preferences in order of distance from stadiums.