Pre meeting updates
Made a change to stadiums so the two memorial stadiums do not seem like the same stadium. Also added utility code to make interacting with voter and candidate objects easier for me. The major update though is the voting.py change to STV, which seems to work for real this time (without messing up future runs or not removing candidates placed in the committee).